+++ MIZDB/settings/defaults.py
@@ -156,3 +156,6 @@ BOOTSTRAP5 = {
 
 # Whether anonymous users can view pages as if they had 'view' permission.
 ANONYMOUS_CAN_VIEW = True
+
+# Log CSRF failures:
+CSRF_FAILURE_VIEW = "dbentry.csrf.csrf_failure"